requirements.txt
numpy
pandas
pulp
pytest-cov
sympy
